dropped assets interface and integrated into scene
