+++ functest/example_custom_renderer.py
@@ -6,7 +6,7 @@
 #       extension: .py
 #       format_name: light
 #       format_version: '1.4'
-#       jupytext_version: 1.2.1
+#       jupytext_version: 1.2.4
 #   kernelspec:
 #     display_name: Python 3
 #     language: python
@@ -28,12 +28,14 @@
 # In this time, we implemented two functions: :cpp:func:`lm::Component::construct` and :cpp:func:`lm::Renderer::render`. :cpp:func:`lm::Component::construct` function provides a type-agnostic way to initialize the instance with JSON values. You want to implement main logic of the renderer inside the :cpp:func:`lm::Renderer::render` function. We will not explain the detail here, but this renderer implements a simple ambient occlusion. As for the usage of APIs, please refer to the :ref:`corresponding references <api_ref>` for detail.
 # -
 
+import lmenv
+env = lmenv.load('.lmenv')
+
 import os
 import numpy as np
 import imageio
 # %matplotlib inline
 import matplotlib.pyplot as plt
-import lmfunctest as ft
 import lightmetrica as lm
 # %load_ext lightmetrica_jupyter
 
@@ -42,7 +44,7 @@ lm.log.init('logger::jupyter')
 lm.progress.init('progress::jupyter')
 lm.info()
 
-lm.comp.loadPlugin(os.path.join(ft.env.bin_path, 'functest_renderer_ao'))
+lm.comp.loadPlugin(os.path.join(env.bin_path, 'functest_renderer_ao'))
 
 # +
 # Film for the rendered image
@@ -61,7 +63,7 @@ lm.asset('camera1', 'camera::pinhole', {
 
 # OBJ model
 lm.asset('obj1', 'model::wavefrontobj', {
-    'path': os.path.join(ft.env.scene_path, 'fireplace_room/fireplace_room.obj')
+    'path': os.path.join(env.scene_path, 'fireplace_room/fireplace_room.obj')
 })
 
 # Camera
