+++ include/lm/debug.h
@@ -18,18 +18,33 @@ LM_NAMESPACE_BEGIN(debug)
 
 /*!
     \brief Poll floating-point value.
+    \param name Query name.
+    \param val Value.
 */
 LM_PUBLIC_API void pollFloat(const std::string& name, Float val);
 
 /*!
     \brief Function being called on polling floating-point values.
+    \param name Query name.
+    \param val Value.
 */
 using OnPollFloatFunc = std::function<void(const std::string& name, Float val)>;
 
 /*!
     \brief Register polling function for floating-point values.
+    \param onPollFloat Callback function.
 */
 LM_PUBLIC_API void regOnPollFloat(const OnPollFloatFunc& onPollFloat);
+
+/*!
+    \brief Attach to debugger.
+
+    \rst
+    Attach the process to Visual Studio Debugger.
+    This function is only available in Windows environment.
+    \endrst
+*/
+LM_PUBLIC_API void attachToDebugger();
 
 /*!
     @}
